refactor: create hook useCreateUserAndNavigate
